Crate systemctl

Crate to manage and monitor services through systemctl
Homepage: https://github.com/gwbres/systemctl

Structs

Unit

Structure to describe a systemd unit

UnitList

Implementation of list generated with systemctl list-unit-files

Enums

AutoStartStatus

AutoStartStatus describes the Unit current state

Doc

Doc describes types of documentation possibly available for a systemd unit

State

State describes a Unit current state

Type

Type describes a Unit declaration Type in systemd

Functions

cat

Invokes systemctl cat on given unit

disable

Disable given unit to start at boot

enable

Enable given unit to start at boot

exists

Returns true if given unit exists, ie., service could be or is actively deployed and manageable by systemd

freeze

Freezes (halts) given unit. This operation might not be feasible.

is_active

Returns true if given unit is actively running

isolate

Isolates given unit, only self and its dependencies are now actively running

list_disabled_services

Returns list of services that are currently declared as disabled

list_enabled_services

Returns list of services that are currently declared as enabled

list_units

Returns a Vector of unit names extracted from systemctl listing.

list_units_full

Returns a Vector of UnitList structs extracted from systemctl listing.

reload

Triggers reload for given unit

reload_or_restart

Triggers reload or restarts given unit

restart

Forces given unit to (re)start

start

Forces given unit to start

status

Returns raw status from systemctl status $unit call

stop

Forces given unit to stop

unfreeze

Unfreezes given unit (recover from halted state). This operation might not be feasible.